Move views to pages directory
