+++ src/components/App/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Login from '../../views/Login/login';
+import Login from '../../pages/Login';
 
 
 function App() {
